+++ esibd/const.py
@@ -65,6 +65,7 @@ FILE_INI = '.ini'
 FILE_H5 = '.h5'
 FILE_PDF = '.pdf'
 FILE_PY = '.py'
+FILE_LOG = '.log'
 
 # other
 UTF8 = 'utf-8'