+++ esibd/devices/ni9263/ni9263.py
@@ -95,3 +95,4 @@ class VoltageController(DeviceController):
         super().closeCommunication()
         # nothing to close
         self.initialized = False
+        self.closing = False