+++ esibd/devices/omnicontrol/omnicontrol.py
@@ -111,3 +111,4 @@ class PressureController(DeviceController):
                 self.port.close()
                 self.port = None
         self.initialized = False
+        self.closing = False