+++ esibd/examples/custom_device/custom_device.py
@@ -223,4 +223,6 @@ class CustomController(DeviceController):
                 # try to close port even if lock could not be acquired! resulting errors should be excepted
                 self.port.close()
                 self.port = None
-        self.initialized = False  # call this last
+        # call thess last
+        self.initialized = False
+        self.closing = False